requirements.txt
numpy
Pillow
matplotlib
streamlit
chromadb==0.5.5
open-clip-torch
beautifulsoup4
selenium==4.23.1
tqdm
